Replace "Oblivious" by "Instruction" and add UI improvements
I did not get permission to re-distribute the Oblivious font which I had
originally used for parts of the UI. The license was very clear: "Do not
distribute without the author's permission.", which, in my opinion,
includes re-distribution embedded in software as well. "Instruction"
clearly allows embedding and also is very beautiful.

Also added some more columns to the UI to make it prettier and more
informative.